Background music added, requires implementation.
Music is written in a way that only works well if looping is smooth.
If not, let me know and I will rewrite parts to make it not need smooth
looping.
I can't seem to get the game to run on my end, so I will need someone
else to implement the music